♻️ refactor:  AuthRepositoryをUserRepositoryに統合
#3
