+++ lib/domain/repository/user/entity/auth_provider.dart
