+++ lib/domain/repository/user/entity/auth_user.freezed.dart
