+++ lib/domain/repository/user/user_repository.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'entity/auth_user.dart';
 import 'entity/user.dart';
 
 /// ユーザーリポジトリプロバイダー
@@ -10,6 +11,18 @@ final userRepositoryProvider = Provider<UserRepository>(
 
 /// ユーザーリポジトリ
 abstract class UserRepository {
+  /// 認証ユーザーを返す
+  AuthUser? getAuthUser();
+
+  /// 匿名認証でログインする
+  Future<AuthUser> loginAnonymously();
+
+  /// ログアウトする
+  Future<void> logout();
+
+  /// 認証ユーザーを削除する
+  Future<void> delete();
+
   /// ユーザーを監視する
   Stream<User?> changes();
