+++ lib/infrastructure/firebase/firestore.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../firebase_auth/auth.dart';
+import 'auth.dart';
 
 /// FirebaseFirestoreプロバイダー
 final firebaseFirestoreProvider = Provider(
