+++ lib/main.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'domain/repository/auth/auth_repository.dart';
 import 'domain/repository/user/user_repository.dart';
-import 'infrastructure/firebase_auth/auth.dart';
-import 'infrastructure/firebase_auth/auth/auth_repository.dart';
-import 'infrastructure/firebase_core/core.dart';
-import 'infrastructure/firebase_firestore/firestore.dart';
-import 'infrastructure/firebase_firestore/user/user_repository.dart';
+import 'infrastructure/firebase/auth.dart';
+import 'infrastructure/firebase/core.dart';
+import 'infrastructure/firebase/firestore.dart';
+import 'infrastructure/firebase/user/user_repository.dart';
 import 'presentation/app.dart';
 import 'util/logger.dart';
 
@@ -24,15 +22,11 @@ Future<void> main() async {
       ],
       overrides: [
         // 各 Repository の上書き
-        authRepositoryProvider.overrideWith(
-          (ref) => FirebaseAuthRepository(
-            auth: ref.watch(firebaseAuthProvider),
-            firebaseUser: ref.watch(firebaseUserProvider).value,
-          ),
-        ),
+
         userRepositoryProvider.overrideWith(
           (ref) {
             final repository = FirebaseUserRepository(
+              auth: ref.watch(firebaseAuthProvider),
               firebaseUser: ref.watch(firebaseUserProvider).value,
               docRef: ref.watch(userDocRefProvider),
             );
